+++ frontend/src/components/UI/GlassmorphismForm.tsx
@@ -1,8 +1,9 @@
-import styles from "../../styles/UI/GlassmorphismForm.module.css";
+import styles from "@/styles/UI/GlassmorphismForm.module.css";
 import Squares from "./Squares";
 
 interface GlassmorphismFormProps extends React.PropsWithChildren {
 	addSquares: boolean;
+	onSubmit?: React.FormEventHandler<HTMLFormElement>;
 }
 
 const GlassmorphismForm: React.FC<GlassmorphismFormProps> = (
@@ -10,7 +11,9 @@ const GlassmorphismForm: React.FC<GlassmorphismFormProps> = (
 ) => {
 	return (
 		<Squares addSquares={props.addSquares}>
-			<form className={styles["glassmorphism-form"]}>{props.children}</form>
+			<form className={styles["glassmorphism-form"]} onSubmit={props.onSubmit}>
+				{props.children}
+			</form>
 		</Squares>
 	);
 };